requirements-2.txt
streamlit>=1.34.0
python-docx>=1.0.0
